Fixes to parser plus extra fuzz example.
The parser had a bunch of silent failures, and it probably still does.  This
patch tightens it up a little bit.  It now checks that the variables mentioned
in a ghostbust annotation do not match the ones in the head of the data type decl.
It also errors out when it finds a type argument it can't handle, instead of
silently dropping it.  (This happened for bang patterns, but now it handles
bang patterns and just ignores the bang.)