Perform dodgy type variable renaming
Perform dodgy renaming of the type variables at the head of the GADT declaration. These should really be fresh names that we know don't appear in any of the constructors, otherwise the code generator will create type representation terms that will lead to name capture.

Basically, we don't want something like this:

> data Map k a where       -- must rename this 'k' and 'a'
>    Tip :: Map k a

close #12