Add example from unordered-containers
This data type now parses correctly, and if we set all variables to synthesise we can run the fuzz tester and get some cases that actually generate code. However, due to an error in `splitTyArgs` we can't actually take those (working) settings and feed them back in, so this example is still mostly useless.