fuzzTest: be careful not to expand this enormous list
To determine the number of possible weakening this was calling length on the resulting cartesian product of definitions. Instead, calculate this number ourselves. At an example, the module Data.Array.Acccelerate.AST results in 3057647616000 possible weakening.

The thinning step works on the (lazily generated) cartesian product, and takes only the first 1024 possibilities.